Run `proto-build` automatically with GitHub Actions (#24)
Adds a `rebuild-protos` GitHub Actions workflow to automatically run
`proto-build` every time a commit is merged/pushed to the `main` branch.

After running `rebuild-protos`, if there are any changed files, it uses
the `peter-evans/create-pull-request` GitHub Action to open a PR with the
changes automatically, thus ensuring `cosmos-sdk-proto` is reproducible
and automating the work of regenerating the protos so it can be kept
out-of-band with other PRs.

If no files are changed, the workflow otherwise completes successfully.

Note that this is using a 3rd party GitHub Action, and exposing the
repo's `GITHUB_TOKEN` to it. However, the action is pinned to the latest
release (`peter-evans/create-pull-request@v3.5.2`) effectively making it
trust-on-first-use (i.e. if future updates to the action introduce
malware, we will not be affected unless we choose to bump versions. If
however the malware is already there, we're vulnerable)